requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
distlib==0.3.8
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
platformdirs==4.1.0
PyJWT==2.8.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
virtualenv==20.25.0
